+++ test/unittests.h
@@ -24,7 +24,7 @@ bool unitTest()
     add(test_lib_cvar());
     add(test_db());
     add(test_osl_language());
-    //add(test_zorder());
+    add(test_zorder());
     add(test_sessions());
     //add(test_count_methods());
